refactor: Session catalina 패키지로 변경
