+++ tomcat/src/main/java/org/apache/catalina/SessionManager.java
@@ -1,20 +1,23 @@
-package org.apache.coyote.http11.common;
+package org.apache.catalina;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class SessionManager {
+public class SessionManager implements Manager {
 
     private static final Map<String, Session> SESSIONS = new HashMap<>();
 
+    @Override
     public void add(final Session session) {
         SESSIONS.put(session.getId(), session);
     }
 
+    @Override
     public Session findSession(final String id) {
         return SESSIONS.get(id);
     }
 
+    @Override
     public void remove(final String id) {
         SESSIONS.remove(id);
     }
